+++ test/linting/mypy.sh
@@ -1,9 +1,10 @@
 #!/bin/bash
 
+base_dir="$(realpath $(dirname $0))"
 root_dir="$(realpath $(dirname $0)/../..)"
 
-for version in 3.7 3.8 3.9 3.10 3.11; do
+for version in 3.9 3.10 3.11 3.12 3.13; do
 	echo -n "checking for Python ${version}..."
-	mypy --python-version=${version} ${root_dir}/src/wsdd.py
+	mypy --config-file="${base_dir}/mypy.ini" --python-version=${version} ${root_dir}/src/wsdd.py
 	echo
 done